Add helpers commands for the "Getting started" docs
There's a guide at https://getkong.org/docs/0.10.x/getting-started/quickstart/
which takes you through adding an api, plugin and consumer. As a
way of playing around with make a little and getting more familiar
with Kong I've added a Makefile from which you should be able to
run `make gettingStarted` and it will go through the stages.

One oddity is that on adding the key-auth plugin we don't get 403
as stated in the documentation. If we restart Kong we do start
getting 401.

Note that I've also changed docker-compose.yml to set
KONG_NGINX_DAEMON=off otherwise the container just exits. I'm not
sure what other side effects this could have.